extract AutoCompletePopover into own component
