+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/AutoCompletePopover/README.md
@@ -1,0 +1,1 @@
+This component is the popover showing the suggestions in the `SingleAutoComplete` and `MultiAutoComplete` field.
