+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/AutoCompletePopover/autoCompletePopover.scss
