+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/AutoCompletePopover/index.js
@@ -1,0 +1,4 @@
+// @flow
+import AutoCompletePopover from './AutoCompletePopover';
+
+export default AutoCompletePopover;
