+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/AutoCompletePopover/suggestion.scss
