+++ src/Sulu/Bundle/AdminBundle/Resources/js/components/AutoCompletePopover/tests/Suggestion.test.js
